UI: uniformisation of all the future events on jqGrid

+++ ui/deleteSample.js
@@ -1,9 +1,9 @@
 "use strict";
 
 ui.deleteSample = function( uisample ) {
-	var i = ui.samples.indexOf( uisample );
-	if ( i >= 0 ) {
+	if ( uisample ) {
+		var i = ui.samples.indexOf( uisample );
 		ui.samples.splice( i, 1 );
+		uisample.delete();
 	}
-	uisample.delete();
 };
